refactor - Car에 distance라는 String 원시값을 Position 이라는 객체로 리팩토링
