+++ src/main/java/racingCar/domain/Position.java
@@ -1,0 +1,18 @@
+package racingCar.domain;
+
+public class Position {
+
+    private String position = "";
+
+    public Position() {
+    }
+
+    public String forward() {
+        position += "-";
+        return position;
+    }
+
+    public String getPosition() {
+        return position;
+    }
+}
