+++ src/main/java/racingCar/exception/InvalidCarNameException.java
@@ -1,4 +1,4 @@
-package racingCar;
+package racingCar.exception;
 
 public class InvalidCarNameException extends RuntimeException {
